+++ src/ScoreCalculation/ResultsForUser.php
@@ -6,6 +6,11 @@ use App\Entity\User;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
+/**
+ * Represents the betting that happend for a single user - so all race of one user for one season are considered here.
+ * This is allows finding out which results were the worst for a single user - and marking them as scores which are
+ * to be discarded in the season overview.
+ */
 class ResultsForUser
 {
     // TODO consider making this configurable in yaml.
@@ -16,11 +21,13 @@ class ResultsForUser
 
     public function __construct(
         private User $user,
-        /** @var Collection<int, RaceScoreCalculator> $allRaceScoreCalculators */
-        Collection &$allRaceScoreCalculators
+        /** @var Collection<int, RaceScoreCalculator> $raceScoreCalculatorsOfUser */
+        Collection &$raceScoreCalculatorsOfUser
     ) {
         $this->raceScoreCalculatorsOfUser = new ArrayCollection();
-        $this->updateRaceScoreCalculatorsDiscardData($allRaceScoreCalculators);
+        $this->updateRaceScoreCalculatorsDiscardData($raceScoreCalculatorsOfUser);
+        // TODO make private variables and set their values in constructor, like with ResultsForRace.
+        // THis entity is for user details. Similiar to race.
     }
 
     public function getRaceScoreCalculatorsOfUser(): Collection
@@ -34,16 +41,11 @@ class ResultsForUser
     }
 
     /**
-     * @param Collection<int, RaceScoreCalculator> $allRaceScoreCalculators
+     * @param Collection<int, RaceScoreCalculator> $raceScoreCalculatorsOfUser
      * @return void
      */
-    private function updateRaceScoreCalculatorsDiscardData(Collection &$allRaceScoreCalculators): void {
-        // Filter all race score calculators
-        $this->raceScoreCalculatorsOfUser = $this->filterRaceScoreCalculatorsByUser($allRaceScoreCalculators);
-
-        if ($this->raceScoreCalculatorsOfUser->count() === 0) {
-            return;
-        }
+    private function updateRaceScoreCalculatorsDiscardData(Collection &$raceScoreCalculatorsOfUser): void {
+        $this->raceScoreCalculatorsOfUser = $raceScoreCalculatorsOfUser;
 
         // Find the discarded race score calculators from the users race score calculators..
         /** @var Collection<int, RaceScoreCalculator> $scoreSortedRaceScoreCalculators */